🦄 refactor: engrave cost

+++ src/RebornStorage.sol
@@ -1,16 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity 0.8.17;
 
-import {IRebornPortal} from "src/interfaces/IRebornPortal.sol";
-import {SafeERC20} from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
-import {SafeERC20Upgradeable} from "@openzeppelin/contracts-upgradeable/token/ERC20/utils/SafeERC20Upgradeable.sol";
-
-import {UUPSUpgradeable} from "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
-import {SafeOwnableUpgradeable} from "@p12/contracts-lib/contracts/access/SafeOwnableUpgradeable.sol";
-
-import {IRebornToken} from "src/interfaces/IRebornToken.sol";
 import {IRebornDefination} from "src/interfaces/IRebornPortal.sol";
-
 import {RBT} from "src/RBT.sol";
 
 contract RebornStorage is IRebornDefination {
@@ -27,7 +18,7 @@ contract RebornStorage is IRebornDefination {
 
     mapping(address => bool) public signers;
 
-    mapping(address => uint16) public rounds;
+    mapping(address => uint32) public rounds;
 
     mapping(uint256 => LifeDetail) public details;
